+++ app/src/main/java/com/practicum/playlistmaker/UI/SettingsActivity.kt
@@ -63,8 +63,7 @@ class SettingsActivity : AppCompatActivity() {
         }
 
         arrowBack.setOnClickListener {
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
+            finish()
         }
 
     }
