+++ app/src/main/java/com/practicum/playlistmaker/appSettings/App.kt
@@ -40,41 +40,7 @@ class App : Application() {
         )
     }
 
-    @RequiresApi(Build.VERSION_CODES.N)
-    fun saveSearchStoryPreference(track: Track) {
-        val storyListPreference: Array<Track>? =
-            if (createTrackListFromJson() == null) arrayOf() else createTrackListFromJson()
-        val storyList: ArrayList<Track> = arrayListOf()
-        if (storyListPreference != null) {
-            for (i in storyListPreference) {
-                storyList.add(i)
-            }
-        }
-        storyList.removeIf { element -> element == track }
-        storyList.add(track)
 
-        if (storyList.size > 10) {
-            storyList.removeAt(0)
-        }
-
-        storyPreference.edit().putString(
-            SEARCH_STORY_KEY, createJsonFromTrackList(
-                storyList
-            )
-        )
-            .apply()
-
-    }
-
-    private fun createTrackListFromJson(): Array<Track>? {
-        val sharedTrackList = storyPreference.getString(SEARCH_STORY_KEY, null)
-        return Gson().fromJson(sharedTrackList, Array<Track>::class.java)
-
-    }
-
-    fun createJsonFromTrackList(trackList: ArrayList<Track>): String {
-        return Gson().toJson(trackList)
-    }
 
 
 }
